# [Enhancement] Improve the profile serialization performance