@wanpengfei-git 2023-08-31
#### [FE Incremental Coverage Report]

:disappointed: **fail** : 5 / 9 (55.56%)




#### file detail

|   |path|covered_line|new_line|coverage|not_covered_line_detail|
|----|----|----|----|----|----|
|:large_blue_circle:|com/starrocks/qe/StmtExecutor.java|0|4|00.00%|[320, 322, 323, 327]|
|:large_blue_circle:|com/starrocks/common/util/RuntimeProfile.java|5|5|100.00%|[]|